requirements.txt
click==8.1.3
gitdb==4.0.10
GitPython==3.1.29
smmap==5.0.0
